// nook-web/src/lib/types/manage.gen.ts
// Code generated by tygo. DO NOT EDIT.
import { ModuleMetadata } from "./module.gen";

//////////
// source: config.go

export interface ManageConfig {
}

//////////
// source: wire.go

export interface GlobalModuleWire {
  id: string;
  metadata: ModuleMetadata;
}
export type GlobalModuleListResponseWire = GlobalModuleWire[];
export interface ModuleWire {
  id: string;
  metadata: ModuleMetadata;
  commands: any /* discord.ApplicationCommandCreate */[];
  enabled: boolean;
}
export interface ModuleWithConfigWire extends ModuleWire {
  command_overwrites: { [key: string]: ModuleCommandOverwriteWire};
  config: Record<string, any> | null;
}
export type ModuleGetResponseWire = ModuleWithConfigWire;
export type ModuleListResponseWire = ModuleWire[];
export interface ModuleConfigureRequestWire {
  enabled: boolean;
  command_overwrites: { [key: string]: ModuleCommandOverwriteWire};
  config: Record<string, any> | null;
}
export type ModuleConfigureResponseWire = ModuleWithConfigWire;
export interface ModuleCommandOverwriteWire {
  disabled?: boolean;
}